refactor: relocate function for the format customer data for registration
